refactor: Extract Participant class
