+++ system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/dataspaceconnector/test/e2e/EndToEndTransferTest.java
@@ -14,103 +14,43 @@
 
 package org.eclipse.dataspaceconnector.test.e2e;
 
+import org.eclipse.dataspaceconnector.common.annotations.EndToEndTest;
 import org.eclipse.dataspaceconnector.junit.launcher.EdcRuntimeExtension;
-import org.eclipse.dataspaceconnector.policy.model.Action;
-import org.eclipse.dataspaceconnector.policy.model.Permission;
-import org.eclipse.dataspaceconnector.policy.model.Policy;
-import org.eclipse.dataspaceconnector.policy.model.PolicyType;
-import org.eclipse.dataspaceconnector.spi.asset.AssetSelectorExpression;
 import org.eclipse.dataspaceconnector.spi.types.domain.DataAddress;
-import org.eclipse.dataspaceconnector.spi.types.domain.catalog.Catalog;
-import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferType;
-import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import java.net.URI;
 import java.time.Duration;
 import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static io.restassured.RestAssured.given;
-import static io.restassured.http.ContentType.JSON;
-import static java.io.File.separator;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
-import static org.eclipse.dataspaceconnector.common.testfixtures.TestUtils.getFreePort;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.ENDPOINT;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.NAME;
 import static org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcessStates.COMPLETED;
 import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.notNullValue;
 
-//@EndToEndTest
+@EndToEndTest
 class EndToEndTransferTest {
 
     private final Duration timeout = Duration.ofSeconds(30);
 
-    private static final URI CONSUMER_CONTROL_PLANE = URI.create("http://localhost:" + getFreePort());
-    private static final URI CONSUMER_CONTROL_PLANE_VALIDATION = URI.create("http://localhost:" + getFreePort() + "/validation");
-    private static final URI CONSUMER_CONTROL_PLANE_DATAPLANE = URI.create("http://localhost:" + getFreePort() + "/dataplane");
-    private static final URI CONSUMER_DATA_PLANE = URI.create("http://localhost:" + getFreePort());
-    private static final URI CONSUMER_DATA_PLANE_CONTROL = URI.create("http://localhost:" + getFreePort() + "/control");
-    private static final URI CONSUMER_DATA_PLANE_PUBLIC = URI.create("http://localhost:" + getFreePort() + "/public");
-    private static final URI CONSUMER_BACKEND_SERVICE = URI.create("http://localhost:" + getFreePort());
-    private static final URI CONSUMER_IDS_API = URI.create("http://localhost:" + getFreePort());
-    private static final URI PROVIDER_CONTROL_PLANE = URI.create("http://localhost:" + getFreePort());
-    private static final URI PROVIDER_CONTROL_PLANE_VALIDATION = URI.create("http://localhost:" + getFreePort() + "/validation");
-    private static final URI PROVIDER_CONTROL_PLANE_DATAPLANE = URI.create("http://localhost:" + getFreePort() + "/dataplane");
-    private static final URI PROVIDER_DATA_PLANE = URI.create("http://localhost:" + getFreePort());
-    private static final URI PROVIDER_DATA_PLANE_CONTROL = URI.create("http://localhost:" + getFreePort() + "/control");
-    private static final URI PROVIDER_DATA_PLANE_PUBLIC = URI.create("http://localhost:" + getFreePort() + "/public");
-    private static final URI PROVIDER_BACKEND_SERVICE = URI.create("http://localhost:" + getFreePort());
-    private static final URI PROVIDER_IDS_API = URI.create("http://localhost:" + getFreePort());
-    private static final String IDS_PATH = "/api/v1/ids";
+    private static final Participant CONSUMER = new Participant();
+    private static final Participant PROVIDER = new Participant();
 
     @RegisterExtension
     static EdcRuntimeExtension consumerControlPlane = new EdcRuntimeExtension(
             ":system-tests:e2e-transfer-test:control-plane",
             "consumer-control-plane",
-            new HashMap<>() {
-                {
-                    put("web.http.port", String.valueOf(CONSUMER_CONTROL_PLANE.getPort()));
-                    put("web.http.path", "/api");
-                    put("web.http.ids.port", String.valueOf(CONSUMER_IDS_API.getPort()));
-                    put("web.http.ids.path", IDS_PATH);
-                    put("web.http.dataplane.port", String.valueOf(CONSUMER_CONTROL_PLANE_DATAPLANE.getPort()));
-                    put("web.http.dataplane.path", CONSUMER_CONTROL_PLANE_DATAPLANE.getPath());
-                    put("web.http.validation.port", String.valueOf(CONSUMER_CONTROL_PLANE_VALIDATION.getPort()));
-                    put("web.http.validation.path", "/validation");
-                    put("edc.vault", resourceAbsolutePath("consumer-vault.properties"));
-                    put("edc.keystore", resourceAbsolutePath("certs/cert.pfx"));
-                    put("edc.keystore.password", "123456");
-                    put("ids.webhook.address", CONSUMER_IDS_API.toString());
-                    put("edc.receiver.http.endpoint", CONSUMER_BACKEND_SERVICE + "/api/service/pull");
-                    put("edc.transfer.dataplane.token.signer.privatekey.alias", "1");
-                    put("edc.public.key.alias", "public-key");
-                    put("edc.transfer.dataplane.sync.endpoint", CONSUMER_DATA_PLANE_PUBLIC.toString());
-                }
-            }
+            CONSUMER.controlPlaneConfiguration()
     );
 
     @RegisterExtension
     static EdcRuntimeExtension consumerDataPlane = new EdcRuntimeExtension(
             ":system-tests:e2e-transfer-test:data-plane",
             "consumer-data-plane",
-            new HashMap<>() {
-                {
-                    put("web.http.port", String.valueOf(CONSUMER_DATA_PLANE.getPort()));
-                    put("web.http.path", "/api");
-                    put("web.http.public.port", String.valueOf(CONSUMER_DATA_PLANE_PUBLIC.getPort()));
-                    put("web.http.public.path", "/public");
-                    put("web.http.control.port", String.valueOf(CONSUMER_DATA_PLANE_CONTROL.getPort()));
-                    put("web.http.control.path", CONSUMER_DATA_PLANE_CONTROL.getPath());
-                    put("edc.controlplane.validation-endpoint", CONSUMER_CONTROL_PLANE_VALIDATION + "/validation");
-                }
-            }
+            CONSUMER.dataPlaneConfiguration()
     );
 
     @RegisterExtension
@@ -119,7 +59,7 @@ class EndToEndTransferTest {
             "consumer-backend-service",
             new HashMap<>() {
                 {
-                    put("web.http.port", String.valueOf(CONSUMER_BACKEND_SERVICE.getPort()));
+                    put("web.http.port", String.valueOf(CONSUMER.backendService().getPort()));
                 }
             }
     );
@@ -128,43 +68,14 @@ class EndToEndTransferTest {
     static EdcRuntimeExtension providerDataPlane = new EdcRuntimeExtension(
             ":system-tests:e2e-transfer-test:data-plane",
             "provider-data-plane",
-            new HashMap<>() {
-                {
-                    put("web.http.port", String.valueOf(PROVIDER_DATA_PLANE.getPort()));
-                    put("web.http.path", "/api");
-                    put("web.http.public.port", String.valueOf(PROVIDER_DATA_PLANE_PUBLIC.getPort()));
-                    put("web.http.public.path", "/public");
-                    put("web.http.control.port", String.valueOf(PROVIDER_DATA_PLANE_CONTROL.getPort()));
-                    put("web.http.control.path", PROVIDER_DATA_PLANE_CONTROL.getPath());
-                    put("edc.controlplane.validation-endpoint", PROVIDER_CONTROL_PLANE_VALIDATION + "/validation");
-                }
-            }
+            PROVIDER.dataPlaneConfiguration()
     );
 
     @RegisterExtension
     static EdcRuntimeExtension providerControlPlane = new EdcRuntimeExtension(
             ":system-tests:e2e-transfer-test:control-plane",
             "provider-control-plane",
-            new HashMap<>() {
-                {
-                    put("web.http.port", String.valueOf(PROVIDER_CONTROL_PLANE.getPort()));
-                    put("web.http.path", "/api");
-                    put("web.http.ids.port", String.valueOf(PROVIDER_IDS_API.getPort()));
-                    put("web.http.ids.path", IDS_PATH);
-                    put("web.http.dataplane.port", String.valueOf(PROVIDER_CONTROL_PLANE_DATAPLANE.getPort()));
-                    put("web.http.dataplane.path", PROVIDER_CONTROL_PLANE_DATAPLANE.getPath());
-                    put("web.http.validation.port", String.valueOf(PROVIDER_CONTROL_PLANE_VALIDATION.getPort()));
-                    put("web.http.validation.path", "/validation");
-                    put("edc.vault", resourceAbsolutePath("provider-vault.properties"));
-                    put("edc.keystore", resourceAbsolutePath("certs/cert.pfx"));
-                    put("edc.keystore.password", "123456");
-                    put("ids.webhook.address", PROVIDER_IDS_API.toString());
-                    put("edc.receiver.http.endpoint", PROVIDER_BACKEND_SERVICE + "/api/service/pull");
-                    put("edc.transfer.dataplane.token.signer.privatekey.alias", "1");
-                    put("edc.public.key.alias", "public-key");
-                    put("edc.transfer.dataplane.sync.endpoint", PROVIDER_DATA_PLANE_PUBLIC.toString());
-                }
-            }
+            PROVIDER.controlPlaneConfiguration()
     );
 
     @RegisterExtension
@@ -173,7 +84,7 @@ class EndToEndTransferTest {
             "provider-backend-service",
             new HashMap<>() {
                 {
-                    put("web.http.port", String.valueOf(PROVIDER_BACKEND_SERVICE.getPort()));
+                    put("web.http.port", String.valueOf(PROVIDER.backendService().getPort()));
                 }
             }
     );
@@ -182,27 +93,62 @@ class EndToEndTransferTest {
     void httpPullDataTransfer() {
         createAssetAndContractDefinitionOnProvider();
 
-        var catalog = getCatalog(PROVIDER_IDS_API, CONSUMER_CONTROL_PLANE);
+        var catalog = CONSUMER.getCatalog(PROVIDER.idsEndpoint());
         assertThat(catalog.getContractOffers()).hasSize(1);
 
         var assetId = catalog.getContractOffers().get(0).getAsset().getId();
-
-        var negotiationId = negotiateContractFor(assetId, CONSUMER_CONTROL_PLANE, PROVIDER_IDS_API);
-
-        var contractAgreementId = getContractAgreementId(negotiationId, CONSUMER_CONTROL_PLANE);
+        var negotiationId = CONSUMER.negotiateContract(assetId, PROVIDER);
+        var contractAgreementId = CONSUMER.getContractAgreementId(negotiationId);
 
         assertThat(contractAgreementId).isNotEmpty();
 
-        var transferProcessId = dataRequest(contractAgreementId, assetId, CONSUMER_CONTROL_PLANE, PROVIDER_IDS_API, sync());
+        var transferProcessId = CONSUMER.dataRequest(contractAgreementId, assetId, PROVIDER, sync());
 
         await().atMost(timeout).untilAsserted(() -> {
-            var state = getTransferProcessState(transferProcessId, CONSUMER_CONTROL_PLANE);
+            var state = CONSUMER.getTransferProcessState(transferProcessId);
             assertThat(state).isEqualTo(COMPLETED.name());
         });
 
         await().atMost(timeout).untilAsserted(() -> {
             given()
-                    .baseUri(CONSUMER_BACKEND_SERVICE.toString())
+                    .baseUri(CONSUMER.backendService().toString())
+                    .when()
+                    .get("/api/service/providerData")
+                    .then()
+                    .statusCode(200)
+                    .body("message", equalTo("some information"));
+        });
+    }
+
+    @Test
+    void httpPushDataTransfer() {
+        PROVIDER.registerDataPlane();
+        createAssetAndContractDefinitionOnProvider();
+
+        var catalog = CONSUMER.getCatalog(PROVIDER.idsEndpoint());
+        assertThat(catalog.getContractOffers()).hasSize(1);
+
+        var assetId = catalog.getContractOffers().get(0).getAsset().getId();
+        var negotiationId = CONSUMER.negotiateContract(assetId, PROVIDER);
+        var contractAgreementId = CONSUMER.getContractAgreementId(negotiationId);
+
+        assertThat(contractAgreementId).isNotEmpty();
+
+        var destination = DataAddress.Builder.newInstance()
+                .type("HttpData")
+                .property(NAME, "data")
+                .property(ENDPOINT, CONSUMER.backendService() + "/api/service")
+                .build();
+        var transferProcessId = CONSUMER.dataRequest(contractAgreementId, assetId, PROVIDER, destination);
+
+        await().atMost(timeout).untilAsserted(() -> {
+            var state = CONSUMER.getTransferProcessState(transferProcessId);
+            assertThat(state).isEqualTo(COMPLETED.name());
+        });
+
+        await().atMost(timeout).untilAsserted(() -> {
+            given()
+                    .baseUri(CONSUMER.backendService().toString())
                     .when()
                     .get("/api/service/providerData")
                     .then()
@@ -213,241 +159,13 @@ class EndToEndTransferTest {
 
     private void createAssetAndContractDefinitionOnProvider() {
         var assetId = "asset-id";
-        createAsset(PROVIDER_CONTROL_PLANE, assetId);
-        var policyId = createPolicy(assetId, PROVIDER_CONTROL_PLANE);
-        createContractDefinition(policyId, PROVIDER_CONTROL_PLANE);
-    }
-
-    private Catalog getCatalog(URI provider, URI instance) {
-        return given()
-                .baseUri(instance.toString())
-                .contentType(JSON)
-                .when()
-                .queryParam("providerUrl", provider + IDS_PATH + "/data")
-                .get("/api/catalog")
-                .then()
-                .statusCode(200)
-                .extract().body().as(Catalog.class);
-    }
-
-    @Test
-    void httpPushDataTransfer() {
-        var assetId = "asset-id";
-        createAsset(PROVIDER_CONTROL_PLANE, assetId);
-        var policyId = createPolicy(assetId, PROVIDER_CONTROL_PLANE);
-        createContractDefinition(policyId, PROVIDER_CONTROL_PLANE);
-
-        registerDataPlane(PROVIDER_CONTROL_PLANE_DATAPLANE, PROVIDER_DATA_PLANE_CONTROL);
-        var negotiationId = negotiateContractFor(assetId, CONSUMER_CONTROL_PLANE, PROVIDER_IDS_API);
-
-        var contractAgreementId = getContractAgreementId(negotiationId, CONSUMER_CONTROL_PLANE);
-
-        assertThat(contractAgreementId).isNotEmpty();
-
-        var destination = DataAddress.Builder.newInstance()
-                .type("HttpData")
-                .property(NAME, "data")
-                .property(ENDPOINT, CONSUMER_BACKEND_SERVICE + "/api/service")
-                .build();
-        var transferProcessId = dataRequest(contractAgreementId, assetId, CONSUMER_CONTROL_PLANE, PROVIDER_IDS_API, destination);
-
-        await().atMost(timeout).untilAsserted(() -> {
-            var state = getTransferProcessState(transferProcessId, CONSUMER_CONTROL_PLANE);
-            assertThat(state).isEqualTo(COMPLETED.name());
-        });
-
-        await().atMost(timeout).untilAsserted(() -> {
-            given()
-                    .baseUri(CONSUMER_BACKEND_SERVICE.toString())
-                    .when()
-                    .get("/api/service/providerData")
-                    .then()
-                    .statusCode(200)
-                    .body("message", equalTo("some information"));
-        });
+        PROVIDER.createAsset(assetId);
+        var policyId = PROVIDER.createPolicy(assetId);
+        PROVIDER.createContractDefinition(policyId);
     }
 
     private DataAddress sync() {
         return DataAddress.Builder.newInstance().type("HttpProxy").build();
     }
 
-    private void registerDataPlane(URI controlPlane, URI dataPlane) {
-        var body = Map.of(
-                "edctype", "dataspaceconnector:dataplaneinstance",
-                "id", UUID.randomUUID().toString(),
-                "url", dataPlane.toString() + "/transfer",
-                "allowedSourceTypes", List.of("HttpData"),
-                "allowedDestTypes", List.of("HttpData")
-        );
-
-        given()
-                .baseUri(controlPlane.toString())
-                .contentType(JSON)
-                .body(body)
-                .when()
-                .post("/instances")
-                .then()
-                .statusCode(204);
-    }
-
-    private String getTransferProcessState(String transferProcessId, URI instance) {
-        return given()
-                .baseUri(instance.toString())
-                .contentType(JSON)
-                .when()
-                .get("/api/transferprocess/{id}/state", transferProcessId)
-                .then()
-                .statusCode(200)
-                .extract().body().asString();
-    }
-
-    private String dataRequest(String contractAgreementId, String assetId, URI instance, URI provider, DataAddress destination) {
-        var request = Map.of(
-                "contractId", contractAgreementId,
-                "assetId", assetId,
-                "connectorId", "provider",
-                "connectorAddress", provider + "/api/v1/ids/data",
-                "protocol", "ids-multipart",
-                "dataDestination", destination,
-                "managedResources", false,
-                "transferType", TransferType.Builder.transferType()
-                        .contentType("application/octet-stream")
-                        .isFinite(true)
-                        .build()
-        );
-
-        return given()
-                .baseUri(instance.toString())
-                .contentType(JSON)
-                .body(request)
-                .when()
-                .post("/api/transferprocess")
-                .then()
-                .statusCode(200)
-                .extract().body().asString();
-    }
-
-    private String getContractAgreementId(String negotiationId, URI instance) {
-        var contractAgreementId = new AtomicReference<String>();
-
-        await().atMost(timeout).untilAsserted(() -> {
-            var result = given()
-                    .baseUri(instance.toString())
-                    .contentType(JSON)
-                    .when()
-                    .get("/api/contractnegotiations/{id}", negotiationId)
-                    .then()
-                    .statusCode(200)
-                    .body("contractAgreementId", notNullValue())
-                    .extract().body().jsonPath().getString("contractAgreementId");
-
-            contractAgreementId.set(result);
-        });
-
-        return contractAgreementId.get();
-    }
-
-    private String negotiateContractFor(String assetId, URI instance, URI provider) {
-        var policy = Policy.Builder.newInstance()
-                .id(UUID.randomUUID().toString())
-                .permission(Permission.Builder.newInstance()
-                        .target(assetId)
-                        .action(Action.Builder.newInstance().type("USE").build())
-                        .build())
-                .type(PolicyType.SET)
-                .build();
-        var request = Map.of(
-                "connectorId", "provider",
-                "connectorAddress", provider + "/api/v1/ids/data",
-                "protocol", "ids-multipart",
-                "offer", Map.of(
-                        "offerId", "1:1",
-                        "assetId", assetId,
-                        "policy", policy
-                )
-        );
-
-        return given()
-                .baseUri(instance.toString())
-                .contentType(JSON)
-                .body(request)
-                .when()
-                .post("/api/contractnegotiations")
-                .then()
-                .statusCode(200)
-                .extract().body().asString();
-    }
-
-    private void createAsset(URI instance, String assetId) {
-        var asset = Map.of(
-                "asset", Map.of(
-                        "properties", Map.of(
-                                "asset:prop:id", assetId,
-                                "asset:prop:name", "asset name",
-                                "asset:prop:contenttype", "text/plain",
-                                "asset:prop:policy-id", "use-eu"
-                        )
-                ),
-                "dataAddress", Map.of(
-                        "properties", Map.of(
-                                "endpoint", PROVIDER_BACKEND_SERVICE + "/api/service",
-                                "name", "data",
-                                "type", "HttpData"
-                        )
-                )
-        );
-
-        given()
-                .baseUri(instance.toString())
-                .contentType(JSON)
-                .body(asset)
-                .when()
-                .post("/api/assets")
-                .then()
-                .statusCode(204);
-    }
-
-    private void createContractDefinition(String policyId, URI instance) {
-        var contractDefinition = Map.of(
-                "id", "1",
-                "accessPolicyId", policyId,
-                "contractPolicyId", policyId,
-                "criteria", AssetSelectorExpression.SELECT_ALL.getCriteria()
-        );
-
-        given()
-                .baseUri(instance.toString())
-                .contentType(JSON)
-                .body(contractDefinition)
-                .when()
-                .post("/api/contractdefinitions")
-                .then()
-                .statusCode(204);
-    }
-
-    private String createPolicy(String assetId, URI instance) {
-        var policy = Policy.Builder.newInstance()
-                .permission(Permission.Builder.newInstance()
-                        .target(assetId)
-                        .action(Action.Builder.newInstance().type("USE").build())
-                        .build())
-                .type(PolicyType.SET)
-                .build();
-
-        given()
-                .baseUri(instance.toString())
-                .contentType(JSON)
-                .body(policy)
-                .when()
-                .post("/api/policies")
-                .then()
-                .statusCode(204);
-
-        return policy.getUid();
-    }
-
-    @NotNull
-    private static String resourceAbsolutePath(String filename) {
-        return System.getProperty("user.dir") + separator + "src" + separator + "test" + separator + "resources" + separator + filename;
-    }
 }
